Fixed bug where query params could not be bound to values.
The QueryString in RequestURI was missing.
Fixed RequestURI recreation in URI2PathAndDontSkip(). Now it works in both
CGI and `serve` modes with QueryString().
Added pager for listing celini in stranici and stranici in root page.
Improved CGI detection and emulation.
Default limit for items in a list is now 10.
Improved default Cfg structure generation.